+++ client/src/main/java/com/swust/client/FileSyncClientHandler.java
@@ -3,6 +3,7 @@ package com.swust.client;
 import com.alibaba.fastjson.JSON;
 import com.swust.common.protocol.Message;
 import com.swust.common.protocol.MessageType;
+import com.swust.common.util.FileUtil;
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.Channel;
 import io.netty.channel.ChannelHandler;
